Corrections bugs viewport, refactoring, ajout du zoom initial + action

+++ core/src/main/kotlin/be.catvert.pc/actions/ResizeAction.kt
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.annotation.JsonCreator
 /**
  * Action permettant de redimensionner un gameObject
  */
-class ResizeAction(@ExposeEditor(min = 1, max = Constants.maxGameObjectSize) var newSize: Size) : Action {
+class ResizeAction(@ExposeEditor(min = 1f, max = Constants.maxGameObjectSize.toFloat()) var newSize: Size) : Action {
     @JsonCreator private constructor() : this(Size(1, 1))
 
     override fun invoke(gameObject: GameObject) {
